% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleantfmotiftable.R
\name{cleantfmotiftable}
\alias{cleantfmotiftable}
\title{Clean TF motif table}
\usage{
cleantfmotiftable(tfs)
}
\arguments{
\item{tfs}{Data frame of JASPAR motif database}
}
\value{
tfs2 which is cleaed TF table
}
\description{
\code{cleantfmotiftable} set the TF table
}
\examples{
# tfs2 <- cleantfmotiftable(tfs = tfs)
}
