% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCARTANet.R
\name{CreateCARTANet}
\alias{CreateCARTANet}
\title{Create CARTA Net}
\usage{
CreateCARTANet(seuratobj, df.genome, target, tfs2, id)
}
\arguments{
\item{seuratobj}{Seurat object}

\item{df.genome}{output of data.frame(seuratobj[\link{"ATAC"}]@annotation)}

\item{target}{target of interest (character)}

\item{tfs2}{tfs2}

\item{id}{Motif ID}

\item{tfs}{Data frame of JASPAR motif database}
}
\value{
tfs2 which is cleaed TF table
}
\description{
\code{CreateCARTANet} Create CARTA Network
}
\examples{
# tfs2 <- cleantfmotiftable(tfs = tfs)



}
