% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ranges.R
\name{filter_ranges}
\alias{filter_ranges}
\title{Filter ranges}
\usage{
filter_ranges(
  seuratobj,
  target,
  ranges.links.final,
  corrtheshold,
  assay_exp,
  meantatachreshold
)
}
\arguments{
\item{seuratobj}{Seurat object}

\item{target}{target of interest (character)}

\item{ranges.links.final}{outputs of make_ranges}

\item{corrtheshold}{Correlation value between RNA count and ATAC peak}

\item{assay_exp}{Assays RNA or SCT}

\item{meantatachreshold}{Minimum value within the maximum value for each cluster of ATAC counts)}
}
\value{
ranges.links.final
}
\description{
\code{filter_ranges} Filter ranges
}
