% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idcheck.R
\name{idcheck}
\alias{idcheck}
\title{ID check}
\usage{
idcheck(tfs2, TF)
}
\arguments{
\item{tfs2}{output of cleantfmotiftable}

\item{TF}{Interest TF}
}
\value{
output of combination TF and ID
}
\description{
\code{idcheck} Check ID of TF motifs. If there are some TF motifs, Select single ID by own.
}
\examples{
# id <- idcheck(tfs2 = tfs2, TF = TF)
}
