% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ranges.R
\name{make_ranges}
\alias{make_ranges}
\title{Make ranges}
\usage{
make_ranges(seuratobj, df.genome, target, tfs2, id)
}
\arguments{
\item{seuratobj}{Seurat object}

\item{df.genome}{output of data.frame(seuratobj[\link{"ATAC"}]@annotation)}

\item{target}{Target of interest}

\item{tfs2}{output of cleantfmotiftable}

\item{id}{output of idcheck or interest of TF ID}
}
\value{
ranges.links.final is peak ranges related TSS and gene expression
}
\description{
\code{make_ranges} Make peak ranges related TSS and gene expression
}
\examples{
# ranges.links.final <- make_ranges(seuratobj = gexatac_merge, df.genome = df.genome,  target = target, tfs2 = tfs2, id = id)
}
