% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketfmotiftable.R, R/tfmotif.R
\name{maketfmotiftable}
\alias{maketfmotiftable}
\title{Make TF motif table}
\usage{
maketfmotiftable(PFMatrixList)

maketfmotiftable(PFMatrixList)
}
\arguments{
\item{PFMatrixList}{PFMatrixList from getMatrixSet output}
}
\value{
tfs

tfs
}
\description{
\code{maketfmotiftable}  Make TF motif table

\code{tfmotif}  Make TF motif table
}
\examples{
# tfs <- maketfmotiftable(PFMatrixList = PFMatrixList)
# tfs <- maketfmotiftable(PFMatrixList = PFMatrixList)
}
