% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_check.R
\name{ROI_check}
\alias{ROI_check}
\title{ROI_check}
\usage{
ROI_check(seuratobj, df.genome, target, tfs, fromTSS)
}
\arguments{
\item{seuratobj}{Seurat object}

\item{target}{target of interest (character)}

\item{tfs}{tfs}

\item{fromTSS}{Which peak is the nth peak above or below from TSS?}
}
\value{
txt file with TF motifs matching
}
\description{
\code{ROI_check} Check of Region Of Interest
}
