% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_highlight.R
\name{visualize_highlight}
\alias{visualize_highlight}
\title{Visualize highlight of CARTA output}
\usage{
visualize_highlight(
  seuratobj,
  df.genome,
  target,
  ranges.links.final,
  scorethreshold,
  exp.assay,
  goi
)
}
\arguments{
\item{seuratobj}{Seurat object}

\item{df.genome}{output of data.frame(seuratobj[\link{"ATAC"}]@annotation)}

\item{target}{target of interest (character)}

\item{ranges.links.final}{outputs of make_ranges and/or filtered}

\item{scorethreshold}{Correlation value between RNA count and ATAC peak}

\item{goi}{Genes to show in violin plot}

\item{assay_exp}{Assays RNA or SCT}
}
\value{
ranges.show
}
\description{
\code{visualize_highlight} Visualize highlight of CARTA output
}
\examples{
# visualize_highlight.tmp <- visualize_highlight(seuratobj = gexatac_merge, df.genome = df.genome,  target = target, ranges.links.final = roi)
}
